+++ pkg/components/component.go
@@ -21,16 +21,14 @@ var (
 )
 
 type Component struct {
-	Name          string
-	Instance      string
-	Settings      manifests.Settings
-	Resources     *ResourceSet
-	State         ComponentLifecycle
-	Defaults      map[string]any
-	Services      *ServiceSet
-	Version       int
-	SetupScript   string
-	CleanupScript string
+	Name      string
+	Instance  string
+	Settings  manifests.Settings
+	Resources *ResourceSet
+	State     ComponentLifecycle
+	Defaults  map[string]any
+	Services  *ServiceSet
+	Version   int
 }
 
 //go:generate stringer -type ComponentLifecycle -trimprefix State
@@ -152,8 +150,6 @@ func (c *Component) ApplyManifest(man *manifests.ComponentManifest) error {
 	for _, s := range secretResources {
 		c.Resources.Set(s)
 	}
-	c.SetupScript = man.Settings.SetupScript
-	c.CleanupScript = man.Settings.CleanupScript
 	return nil
 }
 
@@ -186,10 +182,10 @@ func (c Component) Validate() error {
 	if c.Services == nil {
 		return errors.New("component without services set")
 	}
-	if c.SetupScript != "" && c.CleanupScript == "" {
+	if c.Settings.SetupScript != "" && c.Settings.CleanupScript == "" {
 		return errors.New("component has setup script but no cleanup script")
 	}
-	if c.SetupScript == "" && c.CleanupScript != "" {
+	if c.Settings.SetupScript == "" && c.Settings.CleanupScript != "" {
 		return errors.New("component has cleanup script but no setup script")
 	}
 	return nil